refactor: extract shuffle
