+++ docs/pages/array/copy.mdx
@@ -11,7 +11,7 @@ import { copy } from '@fullstacksjs/toolbox';
 ### Signature
 
 ```typescript copy
-copy(array)
+function copy<T extends any[]>(arr: T): T {}
 ```
 
 ### Examples
