+++ docs/pages/array/shuffle.mdx
@@ -1,20 +1,23 @@
 # shuffle
 
-Gets an array and return a shuffled version of it.
+Return a shuffled version of an array
 
-import:
+### Import
+
 ```typescript copy
-import { shuffle } from '@fullstacksjs/toolbox'
+import { shuffle } from '@fullstacksjs/toolbox';
 ```
 
-usage:
+### Signature
+
 ```typescript copy
-shuffle(array)
+function shuffle<T extends any[]>(arr: T): T {}
 ```
 
-example:
+### Examples
+
 ```typescript copy
-copyArray([])                 // []
-copyArray([1])                // [1]
-copyArray([1, 2, 3 ,4 , 5])   // [2, 1, 3, 4, 5]
+shuffle([])        // []
+shuffle([1])       // [1]
+shuffle([1, 2, 3]) // Who knows?
 ```
