+++ src/array/array.ts
@@ -3,12 +3,6 @@ import { randomInt } from '../number/number.js';
 import type { FilterNullish } from '../types/types.js';
 
 /**
- * Gets an array and return a shuffled version of it
- */
-export const shuffle = <T extends any[]>(arr: T): T =>
-  (arr.slice() as T).sort(() => (Math.random() < 0.5 ? 1 : -1));
-
-/**
  * Returns a random item from an array
  */
 export const getRandom = <T>(arr: T[]) => {
